backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.58.0
anthropic>=0.40.0
supabase>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
httpx>=0.28.0
pypdf>=4.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
